Added AbstractRootAppProxy
